+++ src/example_robot_framework/__init__.py
@@ -3,3 +3,7 @@ from . import utils
 
 from .dummy import DummyLibrary
 from .exceptions import ContextException
+from .self_spinning_node import SelfSpinningNode
+from .example_ros2 import ExampleVelocityNode
+from .turtlesim_ros2 import TurtlesimExampleNode
+from .all_calculations import Calculations
