--- src/ros_robot_framework/__init__.py
@@ -1,5 +1,0 @@
-from . import self_spinning_node
-from .self_spinning_node import SelfSpinningNode
-
-from . import example_ros2
-from .example_ros2 import ExampleVelocityNode
