--- src/turtlesim_robot_framework/__init__.py
@@ -1,5 +1,0 @@
-from . import self_spinning_node
-from .self_spinning_node import SelfSpinningNode
-
-from . import turtlesim_ros2
-from .turtlesim_ros2 import TurtlesimExampleNode
